Change fixture reset behavior
On load, set clone qunit-fixture element and set config.fixture equal to
that. Change the reset function to replace the entire qunit-fixture
element with a deep copy of the clone. Essentially, cache a clean copy of
the fixture node on load and replace subsequent fixture nodes with that.

Added tests to check if properties on the fixture are cleared between
tests.