Fix mis-match between moduleStart and moduleDone events
Commit 4817861c02 fixed a bug where moduleStart was incorreclty
emitted. This broke the test/logs.js suite which was relying
on this bug. In fixing that another bug was uncovered.

Tests that run *after* the done event, cause the module they
are in to be ended after each test (because done doesn't know
there may be more tests so it closes any "open" modules).

However when the next test comes along, it doesn't re-open it
because config.previousModule was the same from before it was
closed. It needs to be reset to make sure moduleStart and
moduleDone are always matching.