Tests: Remove after-done tests
The undocumented feature that re-opened the test suite for the html reporter,
when adding more assertions after test suite has finished running tests,
was removed. This only worked in the html reporter.

The callback interface never supported this properly and reporters such as
node-qunit, grunt-contrib-qunit and TestSwam never supported this either.

If this causes trouble, be sure to verify QUnit does not start until after
your tests are loaded (maybe using config.autostart).