Core: Detail modules and tests names in the logging callbacks
QUnit.begin and QUnit.done now send the callback parameter with the
modules property, which now is an Array of objects each one with a
name property representing the module name and a tests property
with an Array of its respective tests names.

moduleDone and moduleStart also send in the callback parameter a
tests property containing an Array of its respective tests names.