Core: Equivalency for desdendants of null constructors
Fixes #851
Closes #856

`Object.create(null);` is costly. A factory object is being used
to mitigate this, setting its prototype with a `constructor`
property described with a null value. This object instances
has dramatically reduced allocation costs.

It is important to allow assert.deepEqual to verify the equivalency
of literal and regular objects with those objects where their
constructor properties with a null value inherit from their prototype.