HTML Reporter: Adds indicator for filtered test
Fixes #528
Closes #876

A new reporter badge will show on the header when QUnit is running
filtered tests through `QUnit.config.testId` API or from the url
parameters triggered after clickin the rerun button.

The badge will tell which tests it is running using the given hash,
followed by a link to clear the filters.