Assert: Implement Assert#pushResult
Fixes gh-827
Closes gh-920

The new pushResult method replaces Assert#push and works in the same way, but
with single parameter to fetch all the data .push received as individual
parameters.

The current assertions already use the new method, while .push is still present,
and it's now a deprecated method, expected to be removed on the next major
version, but updated to map to the newly added method.