Build: Remove browserstack-runner from CI tasks
The new updated version of browserstack-runner (0.4.x) is returning a
constant timeout failure on the amd tests.

Instead of reverting browserstack-runner back to the previous and
deprecated version (0.3.x), we should investigate the current error
and find a proper fix. The old version is expected to stop working
soon, as Browserstack will maintain compatibility only to the new
version.

Ref gh-959