HTML Reporter: - stop button ui
I placed the abort button in the `qunit-testresults` area and made
sure to remove it once the tests were done. I also add some styling
to tests that were not ran because the tests were aborted. I chose
`orange` to stand but it should probably be changed to something
more inline with qunits color pallete.