Reporter: move urlparams import from reporter to html
80a14ee5ac49c7f0981884a6cbb5786dea3a9c8b implemented Rollup.
Reporter is importing `html` before `urlparams`, but `urlparams` needs
to be loaded before `html` as it was in the previous commit.

Tried reordering the imports within the file, but running the build task
seemed to not change the output. By moving the `urlparams` import
into `html`, the built output is correct.

Fixes #1082