Docs: Fix Jekyll build (remove conflicting file/dir names)
Follows-up a7ccfcb.

Having an entry for both "foo" and "foo/" is supported, however
a problem happens if the "foo" part contains one or more dots.

Normally, "foo" creates `_site/foo.html` and
"foo/" creates `_site/foo/index.html`. However, if the name
contains a dot, then the file name won't have ".html" appended.

This causes the Jekyll build to fail because it cannot, for example
create "_site/QUnit.module" as both a file and a directory.

```
jekyll 3.4.3 | Error:
> File exists @ dir_s_mkdir - qunit/docs/_site/QUnit.module
```

Rather than trying to find a way to enforce `.html` in the filename,
I found out that the Jekyll plugin actually handles redirects for
slashless variants as well, so we only need the versions with a slash
and everything will continue to work fine.