Build: Remove Browserstack
Browserstack has long been an "allow failure" and we haven't really
been paying much attention to it. At this point, it is not providing
much value. Additionally, some of the packages we rely on to make it
work have unfixed unvulnerabilites.

Given cross-browser compatibility for JS is much better than in the
past, I don't think we need to replace Browserstack at this point in
time. We can revisit if it becomes an issue in the future.