Report full stack trace for uncaught errors if possible
Currently, when QUnit catches an uncaught error in its window.onerror
handler, it reports it with a source of `fileName + ":" + lineNumber`.

While it's not standard, most modern browsers include an errorObj
argument to window.onerror, which can be used to get a more complete
stacktrace
([source](https://blog.sentry.io/2016/01/04/client-javascript-reporting-window-onerror)).
This change modifies the HTML reporter and the `QUnit.onError` method to
do that.