Build: Add --connect-port for Gruntfile to override 8000
When something is locally running on port 8000 already, the
tests consistently fail. For now, spread the approach we used
for livereload and allow it to be specified by the user.

This is a pretty poor user experience, but until we find a way
to not need, or do differently, the urls in grunt-contrib-qunit,
we need to know it synchronously while calling grunt.initConfig,
which basically leaves manual CLI options as the only option.