Build: Add macOS and Windows to the Travis CI test matrix (#1360)
Update grunt-contrib-qunit to 3.1.0, because grunt-contrib-qunit 3.0.1
uses a pinned (older) version of puppeteer. The older puppteer used an
older version of Chrome, which is incompatible with Windows Server 2016,
which is what Travis CI offers. AppVeyor was using Windows Server 2012,
whih was not affected by that Chrome bug.

Ref https://github.com/gruntjs/grunt-contrib-qunit/issues/154.

For Windows, the build script is limited to `test:main`. This is the
same as it was before with AppVeyor, because the CLI tests are not passing
on Windows currently. Ref https://github.com/qunitjs/qunit/issues/1359.

Rename bin/qunit to bin/qunit.js, because otherwise ESLint fails:

> Expected linebreaks to be 'LF' but found 'CRLF'  linebreak-style

In theory, it should be possible to add an entry to `.gitattributes`
that will apply to bin/qunit what we apply to *.js, but I wasn't
able to find a way.

Fixes #1347.