Assert: Report RegExp/Error as strings from rejects()/throws()
Report the `actual` (Error object) as a string.
And report an `expected` RegExp or Error object also in its
string form.

Currently, they were reported as their objects, which in the
generic js-reporters module was just printed as an empty object
without any properties because instances of RegExp and instances
of Error both have no own properties that are enumerable.

Also fix errorString() which oddly had logic for passing a plain
object, but didn't

Fixes #1333.