Build: Limit Algolia job to direct pushes
Follows-up 2486b22829d5 where I added the Algolia job but wrongly
assumed that "branch = master" would evaluate to false on pull
requests that run on code outside the repository. While it does
reflect the source branch for local branches, it evaluates to
the target branch for remote pull requests.

The good news is that the job simply failed since the secure token
is not exposed to pull requests.