Tests: Fix "source location" test in IE 11
We are able to get a stacktrace synchronously via QUnit.stack()
in IE 11 by throw-catching the Error object directly.

However, while we do that for global error reporting, we don't use
this code path in the HTML reporter (for performance reasons).
We only store the Error object in /src/core/test.js and the HTML
reporter lazily gets the stack trace. In other words, we call
extractStacktrace() but we don't call sourceFromStacktrace() on
passing tests unless we're in the HTML reporter.

Long story short, this means we don't suppor traces in IE 11
except for reporting failing tests and global errors which is
how it has been for a few years. The only "issue" is that our
unit test for this feature used the wrong condition, causing it
to expect to see a trace in IE 11 and thus failing.