Tests: Make the memory-leak even more robust
It only passing some of the time on the latest Node 14 build, confirmed
locally on macOS and also seen at Travis CI on Linux.

It appears V8 now realizes destroy() is a no-op and sometimes inlines
the code or otherwise makes it no longer appear in the heap.

Instead of trying to fool V8 with no-op methods in afterEach(), make
explicitly remain in memory using a regular Set, and then clear it
from there when we're ready for it and calling GC.