Build: Simplify ESLint config and declare "tiny-lr" dev dependency
Let eslint iterate the full directory so that we don't miss out
on any files. Use "ignorePatterns" for any artefacts and such
that we don't want to lint.

Also unify the various configs for Node.js scripts together
in a single override in the `.eslintrc.json` file.

This revealed a problem:

> /qunit/build/tasks/livereload.js
>  2:27  error  "tiny-lr" is extraneous  node/no-extraneous-require

This package was brought in via grunt-contrib-watch, but also used
by our livereload code directly and should have been specified.
This is now fixed.