Build: Remove 'commitplease' hook in favour of contribs doc
I've had to manually disable (and subsequently undo) this in
package.json everytime I run 'npm i'  or 'npm ci'.

A secure dev environment must not allow npm scripts to write to
.git/hooks. I use such an environment, and commitplease is
incompatible with that.

We could register it as part of npm-test or something else, but
in practice it seems more often in the way than not. Especially
as it is not very compatible with a PR-workflow where you might
have fixup commits. I personally don't prefer that workflow, but
it's what works best within GitHub (today). If that gets better,
and we can assume each commit stands on its own, we could add this
back as part of npm-test and/or in CI.

For now, just remove it and I'll take it on us as team to edit
this as needed when performing the merge operation, which GitHub
makes easy to do with it prompting to edit the commit message at
that time.

Also:
- Update rollup dev dependency to address deprecation warning.