Tests: Fix new Dump "function args" test to pass in IE 11
Follows-up d752335, which added the "function args" test case
to the Dump suite. This test accidentally relied on magic names
given to an unnamed function (the function has no name, but
in latest Firefox and Chrome, gets named "f" based on the variable).

This test did not pass in IE 11. Make the test explicitly cover
both ways (named and unnamed) separately.