Tests: Update internal fixture to accomodate Node 15 source map traces
Follows-up e295106, which left one test case still using regexes.
Upon changing that to use the new normalised tracing, I found that
it failed on Node 15 because the format of source map traces changed
between Node 14 and Node 15.

This doesn't matter for QUnit in general, since it just prints
whatever trace was given to it by the runtime, regardless of
formatting, but for our test cases to pass when run against Node 15
we do need to have two test cases or normalise across them. I did the
latter.