Build: Rebuild package-lock.json to bump some indirect dev deps
Also removing Gemfile.lock. This repo mainly uses GitHub Pages, so
it wasn't authoritative anyway. It is useful to ensure consistency
between contributors, but we only have one additional package and
that one's pinned with fairly good sub deps that follow semver.

Not worth the small chance benefit that maybe they'll break semver
and we'll have to figure out what the last working version was to pin
or something like that. And to some extent, being locked is also a
problem because it meant local dev wasn't representative of what we'd
deploy.