Core: Make `tap` and `console` reporters generally available
This change makes the reporters, from the js-reporters package,
generally available through the primary distribution, instead of being
limited to the QUnit CLI.

The built-in reporters are exposed via `QUnit.reporters`.

This is a major step toward truly supporting the TAP ecosystem, and
popular runners such as Airtap for browser testing, and the node-tap
CLI (why not). There is still more work to be done, though, such as
making the library more friendly to require'ing, and further
decoupling of our HTML Runner/Reporter. As such, this remains
experimental as of yet. It can be tried out by calling
`QUnit.reporters.tap.init(QUnit)` between importing QUnit and importing
your tests, e.g. from an inline script or setup script.

This overall direction is outlined in
https://github.com/js-reporters/js-reporters/issues/133, and as such
apart from bridging older versions, the js-reporters package will not
be as actively developed going forward. I'm copying the reporters we
developed there into QUnit repository for further developement here,
without the indirection of that package.

Co-authored-by: Florentin Simion <florentins@gmail.com>
Co-authored-by: Franziska Carstens <franziska.carstens@wwu.de>
Co-authored-by: Martin Olsson <martin@minimum.se>
Co-authored-by: Robert Jackson <me@rwjblue.com>
Co-authored-by: Timo Tijhof <krinklemail@gmail.com>
Co-authored-by: Trent Willis <trentmwillis@gmail.com>
Co-authored-by: Zachary Mulgrew <zackthehuman@users.noreply.github.com>
Co-authored-by: jeberger <jeberger@free.fr>