Core: Fix syntax error for 'let' from kleur.js due to Babel bug
Follows-up f8948c9, which broke the build for Safari 9.1 and IE 9,
which do not support ES6 syntax such as 'let', which kleur.js uses.

It's not clear to how why or how it matters whether babelrc is
configured via a separate file vs not, and even more so since we
do use let/const in our own code and that gets compiled just fine.

Also fix an older IE9 bug (don't recall since when), which is that
while conditional use of `console` is fine, once you open the IE dev
tools, the `console` object does exist, but then it still doesn't
have a `bind()` method, despite generally supporting Function#bind
(it is undefined on host methods, but defined on "regular" functions,
fixed in IE10). Make this work as well while at it.