Build: Fix coveralls branch parsing
Since the move to GitHub Actions, the Coveralls service was no longer
able to translate references such as "refs/heads/main" and correlate
them to branch names as "main".

In chatting with their excellent support, I understand now that this
was because Travis CI was doing things in a non-standard way, and
thus they had an excemption for submissions from Travis CI, but
this marker became sticky for some reason despite our (current) builds
not indicating anything being related to Travis.

Specifying `COVERALLS_SERVICE_NAME` should have sufficed to opt-out
of that workaround, but it seems that didn't do the trick. It needs
an explicit/override for the job ID as well. I'm hoping this can
get upstreamed into node-coveralls, but this will do for now.

Fixes https://github.com/qunitjs/qunit/issues/1602.