Build: Update browserstack-runner to fix `QUnit.todo` support
It is hard to believe, but we actually never had a `QUnit.todo()`
test as part of the main index.html test suite that we run in browsers.

There were secondary tests that only run in Headless Chromium via
grunt-contrib-qunit, and CLI tests, but none that we run cross browser.

In commit 47576c16f4b19 I added test/main/each.js to index.html, not
realizing it was the first such case, and this broke the CI build for
all browsers.

* js-reporters 2.1 adds support for QUnit.todo in the adapter
  so that it correctly emits the `runEnd` event to count these toward
  `testCounts.todo` instead of toward `testCounts.failed`.
  https://github.com/js-reporters/js-reporters/pull/140.

* browserstack-runner 0.9.5-qunitjs.1 updates its copy of js-reporters
  to this version, and also updates its error formatter to not print
  details for error objects from non-failed tests.

Fixes https://github.com/qunitjs/qunit/issues/1622.