HTML Reporter: Add "Rerun failed tests" link
Append a "Rerun failed tests" link to the header.

In order for the link color to be on-brand, apply the same color
as for the "Rerun" link in a passing test, since a passing test
has the same background color as the header.

Fixes https://github.com/qunitjs/qunit/issues/1474.