Docs: Improve explanation of `assert.async()` with count param
Use the shorter `count` name in the docs, matching the source code.

Correct that this is the number of "calls", not "callbacks".

Remove implication that this marks the test as "done", rather it marks
the async pause as done. In phrasing, simply say "resume" instead.

Remove mention of "invoking specified number of times" from the intro
sentence since this is rarely used and might be confusing. Bring it
up in the later sentences instead.