Build: Add copying of CSS to "build:dev" task
Use node-watch to listen for src/qunit.css, and copy it over with
the same preprocess() applied as Rollup and Grunt.

Adjust log messages to differentiate between `[rollup]` and `[css]`.

Remove seemingly redundant `process.on("exit")` listener for cleaning
up the watchers since this is afaik only when terminating the process
manually which happens just fine (and quietly) with and without this.

Closes https://github.com/qunitjs/qunit/issues/1581.