HTML Reporter: Internally use `runStart` and `runEnd` events
Switch from `QUnit.begin()` to `QUnit.done()` to the `runStart` and
`runEnd` events, as well as the data these provide, such as the
test counts and runtime.

As part of making the interface less cluttered, I'm proposing we
remove the runtime and status-specific counts from the realtime
progress text, as these were rapidly changing and are hard to read.

These changes, together, would allow removal of a number of things:

* Remove internal dependency on Test class (in favour of runStart data).
* Remove local counting of test by status (in favour of runEnd data).
* Remove local time measuring.

I've moved the progress line from being the third line in the
testresult-display element, to be the first line. This way, the number
of completed tests will be in the same place as where it is displayed
when the run has ended.

I've also removed the line break between "Running" and the name of
the currently running tests. This way the testresult display will
generally have the same visual height during and after the test run.
(Previously, there was jump ump after the test run, since there are
only two lines in the final state, but with the linebreak we had
three lines when tests are running.)

Ref https://github.com/qunitjs/qunit/issues/1486.