Tests: Enable all test suites in CI
Several of the HTML test suites were previously omitted from CI,
as they either couldn't pass on PhantomJS, or were simply forgotten.

* Add test/headless.html. Also updated to use `QUnit.on()`, and to hide
  debug messages when run in CI to avoid adding this noise.
* Add test/module-filter.html
* Add test/performance-mark.html. Ref https://github.com/qunitjs/qunit/pull/1319.
* Add test/webWorker.html. Ref https://github.com/qunitjs/qunit/pull/1171.
* Remove test/each.html. Duplicate of test/index.html.
* Remove test/stack-errors.html. Redundant with other tests for "noglobals"
  and "onUncaughtError" outside test context etc.
* Improve test/sandboxed-iframe.html, which was passing CI
  but failing manually, due to undefined `__grunt_contrib_qunit__`. We
  now define a fallback locally to ease debugging.

Added to test-on-node: test/module-skip.js, test/module-todo.js. These
were disabled due to use of hacky `QUnit.done()`, but that was recently
fixed with 78bda44791.

Also add various main tests to test/webWorker.js, which had not been
updated in a while. The next commit will add a meta CLI test that
ensures that these are all in sync to avoid staleness in the future.

Ref https://github.com/qunitjs/qunit/issues/1511.