Tests: Remove remaining use of assert within pushResult overrides
These are structurally less stable due to relying on the callback
being invoked for the assertion to take place, and can be difficult
to debug as such.

Also, make sure any proxy calls actually pass `thisValue` along so
as to not rely on legacy `config.current` fallbacks in the
Assert#pushResult method.